requirements.txt
# MANDATORY: Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# MANDATORY: Database
sqlalchemy>=2.0.0
alembic>=1.14.0

# MANDATORY: SQLite async driver (dev) — use asyncpg for PostgreSQL in prod
aiosqlite>=0.20.0

# MANDATORY: Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.7.0
pydantic[email]>=2.10.0

# MANDATORY: Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

# MANDATORY: HTTP client (for external calls if needed)
httpx>=0.28.0

# MANDATORY: Structured logging
python-json-logger>=3.2.0

# MANDATORY: Python multipart for file uploads
python-multipart>=0.0.9

# Testing (optional but included)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
